added comments fixed clap help and other minor bugs
